ST-926 Fix build issues
