+++ src/main/java/uk/gov/hmcts/reform/cosapi/edgecase/event/cic/CicUpdateCaseEvent.java
@@ -1,12 +1,10 @@
-package uk.gov.hmcts.reform.cosapi.edgecase.event.mh;
+package uk.gov.hmcts.reform.cosapi.edgecase.event.cic;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
-import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.ConfigBuilder;
-import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.reform.cosapi.common.config.AppsConfig;
 import uk.gov.hmcts.reform.cosapi.constants.CommonConstants;
 import uk.gov.hmcts.reform.cosapi.edgecase.model.CaseData;
@@ -15,36 +13,27 @@ import uk.gov.hmcts.reform.cosapi.edgecase.model.UserRole;
 import uk.gov.hmcts.reform.cosapi.util.AppsUtil;
 
 import static uk.gov.hmcts.reform.cosapi.edgecase.model.UserRole.CITIZEN;
+import static uk.gov.hmcts.reform.cosapi.edgecase.model.UserRole.CREATOR;
 import static uk.gov.hmcts.reform.cosapi.edgecase.model.access.Permissions.CREATE_READ_UPDATE;
 
 @Component
 @Slf4j
-public class SubmitCaseEvent implements CCDConfig<CaseData, State, UserRole>  {
+public class CicUpdateCaseEvent implements CCDConfig<CaseData, State, UserRole>  {
 
     @Autowired
     AppsConfig appsConfig;
 
     @Override
     public void configure(final ConfigBuilder<CaseData, State, UserRole> configBuilder) {
-
         configBuilder
-            .event(AppsUtil.getExactAppsDetailsByCaseType(appsConfig, CommonConstants.ST_MH_CASE_TYPE).getEventIds()
-                       .getSubmitEvent())
-            .forStates(State.DRAFT)
-            .name("Applicant Submitting the case (mh)")
-            .description("The applicant confirms SOT (mh)")
+            .event(AppsUtil.getExactAppsDetailsByCaseType(appsConfig, CommonConstants.ST_CIC_CASE_TYPE).getEventIds()
+                       .getUpdateEvent())
+            .forStates(State.DRAFT, State.SUBMITTED)
+            .name("Edge case (cic)")
+            .description("Edge case application update (cic)")
             .retries(120, 120)
             .grant(CREATE_READ_UPDATE, CITIZEN)
-            .aboutToSubmitCallback(this::aboutToSubmit);
-    }
-
-    public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(CaseDetails<CaseData, State> details,
-                                                                       CaseDetails<CaseData, State> beforeDetails) {
-
-        return AboutToStartOrSubmitResponse.<CaseData, State>builder()
-            .data(details.getData())
-            .state(details.getState())
-            .build();
+            .grant(CREATE_READ_UPDATE, CREATOR);
     }
 
 }
