+++ src/main/java/uk/gov/hmcts/reform/cosapi/edgecase/event/mh/MhSubmitCaseEvent.java
@@ -4,8 +4,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
+import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.ConfigBuilder;
-import uk.gov.hmcts.reform.cosapi.common.AddSystemUpdateRole;
+import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.reform.cosapi.common.config.AppsConfig;
 import uk.gov.hmcts.reform.cosapi.constants.CommonConstants;
 import uk.gov.hmcts.reform.cosapi.edgecase.model.CaseData;
@@ -13,37 +14,37 @@ import uk.gov.hmcts.reform.cosapi.edgecase.model.State;
 import uk.gov.hmcts.reform.cosapi.edgecase.model.UserRole;
 import uk.gov.hmcts.reform.cosapi.util.AppsUtil;
 
-import java.util.ArrayList;
-
-import static uk.gov.hmcts.reform.cosapi.edgecase.model.State.DRAFT;
 import static uk.gov.hmcts.reform.cosapi.edgecase.model.UserRole.CITIZEN;
 import static uk.gov.hmcts.reform.cosapi.edgecase.model.access.Permissions.CREATE_READ_UPDATE;
 
 @Component
 @Slf4j
-public class CreateCaseEvent implements CCDConfig<CaseData, State, UserRole>  {
-
-    @Autowired
-    private AddSystemUpdateRole addSystemUpdateRole;
+public class MhSubmitCaseEvent implements CCDConfig<CaseData, State, UserRole>  {
 
     @Autowired
     AppsConfig appsConfig;
 
     @Override
     public void configure(final ConfigBuilder<CaseData, State, UserRole> configBuilder) {
-        var defaultRoles = new ArrayList<UserRole>();
-        defaultRoles.add(CITIZEN);
-
-        var updatedRoles = addSystemUpdateRole.addIfConfiguredForEnvironment(defaultRoles);
 
         configBuilder
             .event(AppsUtil.getExactAppsDetailsByCaseType(appsConfig, CommonConstants.ST_MH_CASE_TYPE).getEventIds()
-                       .getCreateEvent())
-            .initialState(DRAFT)
-            .name("Create edge case draft case (mh)")
-            .description("Apply for edge case (mh)")
-            .grant(CREATE_READ_UPDATE, updatedRoles.toArray(UserRole[]::new))
-            .retries(120, 120);
+                       .getSubmitEvent())
+            .forStates(State.DRAFT)
+            .name("Applicant Submitting the case (mh)")
+            .description("The applicant confirms SOT (mh)")
+            .retries(120, 120)
+            .grant(CREATE_READ_UPDATE, CITIZEN)
+            .aboutToSubmitCallback(this::aboutToSubmit);
+    }
+
+    public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(CaseDetails<CaseData, State> details,
+                                                                       CaseDetails<CaseData, State> beforeDetails) {
+
+        return AboutToStartOrSubmitResponse.<CaseData, State>builder()
+            .data(details.getData())
+            .state(details.getState())
+            .build();
     }
 
 }
